Move serde related functions to a separate module
